users

+++ src/main/java/com/example/dwspractica/RatingController.java
@@ -14,6 +14,8 @@ public class RatingController {
     RatingService ratingService;
     @Autowired
     GameService gameService;
+    @Autowired
+    UserService userService;
 
 
     //We can see a game's ratings. If it doesn't exist, we'll see an error
@@ -44,18 +46,24 @@ public class RatingController {
 
     //We are able to create ratings, adding a title, comments and a rating with stars
     @PostMapping("/game/{idGame}/newRating")
-    public String newRating(@RequestParam String title, @RequestParam String comment, @RequestParam int stars, @PathVariable int idGame) {
-        Rating aux = new Rating(stars, title, comment);
+    public String newRating(@RequestParam String title, @RequestParam String comment, @RequestParam int stars, @PathVariable int idGame, HttpServletRequest request) {
+        Rating aux = new Rating(stars, title, comment, userService.getIdFromName(request.getUserPrincipal().getName()));
         ratingService.addRating(idGame, aux);
         return "CreatedRating";
     }
 
     @GetMapping("/game/{idGame}/deleteRating/{id}")
-    public String deleteRating(Model model, @PathVariable int idGame, @PathVariable int id) {
+    public String deleteRating(@PathVariable int idGame, @PathVariable int id, HttpServletRequest request) {
         if (gameService.containsGame(idGame)){
             if (ratingService.containsRating(idGame, id)){
-                ratingService.deleteRating(idGame, id);
-                return "DeletedRating";
+                Rating aux= ratingService.getRatings(idGame, id);
+                int iaux=userService.getIdFromName(request.getUserPrincipal().getName());
+                if (request.isUserInRole("ADMIN")||aux.getId_ur()==iaux){
+                    ratingService.deleteRating(idGame, id);
+                    return "DeletedRating";
+                }else{
+                    return "error/401";
+                }
             }else{
                 return "error/404";
             }
